+++ src/components/Form/Form.component.tsx
@@ -48,7 +48,7 @@ const defaultFormProps: FormDefaultProps<object> = {
 /**
  * A flexible yet expressive form renderer.
  *
- * @version 0.0.6
+ * @version 0.0.7
  *
  * @template {object} TData - The type of the form data
  *