+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- `v0.0.7` Form component
+  - Pass form edit state to `layout.hidden()` function
+
+### Fixed
+
+- `v0.0.3` Action component
+  - Stop event propagation
+
+- `v0.0.2` ContextMenuItem interal component
+  - Stop event propagation
+
 ## 0.32.2 - 2024-09-18
 ### Changed
 - `v0.0.8` ComboBoxButton component